perf(管理程序):优化异步方法和分离功能
WindowsServiceAgentManager管理程序
对代码进行了优化和分离，便于开发管理。
1.优化了加载列表时的异步算法，将曾经的伪异步算法替换为了真异步算法
2.将服务事件从主窗口类中分离为独立的服务事件类
3.将日志中的路径设置为私有静态全局变量
